+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - More compression and decompression targets.
 - New `lines` option for archive/unarchive processors.
 - New `encode` and `decode` processors.
+- New `period_ms` field for the `batch` processor.
 
 ## 0.15.4 - 2018-07-04